feat: Move Search and Self Repository to new folder structure
