+++ src/script/components/InputBar/InputBar.tsx
@@ -40,6 +40,7 @@ import {MessageRepository, OutgoingQuote} from 'Repositories/conversation/Messag
 import {EventRepository} from 'Repositories/event/EventRepository';
 import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
 import {PROPERTIES_TYPE} from 'Repositories/properties/PropertiesType';
+import {SearchRepository} from 'Repositories/search/SearchRepository';
 import {useUserPropertyValue} from 'src/script/hooks/useUserProperty';
 import {CONVERSATION_TYPING_INDICATOR_MODE} from 'src/script/user/TypingIndicatorMode';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
@@ -67,7 +68,6 @@ import {MentionEntity} from '../../message/MentionEntity';
 import {MessageHasher} from '../../message/MessageHasher';
 import {QuoteEntity} from '../../message/QuoteEntity';
 import {useAppMainState} from '../../page/state';
-import {SearchRepository} from '../../search/SearchRepository';
 import {StorageRepository} from '../../storage';
 import {TeamState} from '../../team/TeamState';
