+++ src/script/components/Modals/LegalHoldModal/LegalHoldModal.tsx
@@ -33,13 +33,13 @@ import {ClientRepository} from 'Repositories/client';
 import {ConversationRepository} from 'Repositories/conversation/ConversationRepository';
 import {MessageRepository} from 'Repositories/conversation/MessageRepository';
 import {CryptographyRepository} from 'Repositories/cryptography/CryptographyRepository';
+import {SearchRepository} from 'Repositories/search/SearchRepository';
 import {handleEnterDown} from 'Util/KeyboardUtil';
 import {t} from 'Util/LocalizerUtil';
 
 import {useLegalHoldModalState} from './LegalHoldModal.state';
 
 import {User} from '../../../entity/User';
-import {SearchRepository} from '../../../search/SearchRepository';
 import {TeamRepository} from '../../../team/TeamRepository';
 
 const DISABLE_SUBMIT_TEXT_LENGTH = 1;
