+++ src/script/components/UserSearchableList/UserSearchableList.tsx
@@ -25,13 +25,13 @@ import {debounce} from 'underscore';
 
 import {UserList} from 'Components/UserList';
 import {ConversationState} from 'Repositories/conversation/ConversationState';
+import {SearchRepository} from 'Repositories/search/SearchRepository';
 import {partition} from 'Util/ArrayUtil';
 import {t} from 'Util/LocalizerUtil';
 import {matchQualifiedIds} from 'Util/QualifiedId';
 import {sortByPriority} from 'Util/StringUtil';
 
 import type {User} from '../../entity/User';
-import {SearchRepository} from '../../search/SearchRepository';
 import type {TeamRepository} from '../../team/TeamRepository';
 import {TeamState} from '../../team/TeamState';
