+++ src/script/page/RightSidebar/ConversationParticipants/ConversationParticipants.tsx
@@ -23,6 +23,7 @@ import {FadingScrollbar} from 'Components/FadingScrollbar';
 import {SearchInput} from 'Components/SearchInput';
 import {UserSearchableList} from 'Components/UserSearchableList';
 import {ConversationRepository} from 'Repositories/conversation/ConversationRepository';
+import {SearchRepository} from 'Repositories/search/SearchRepository';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {t} from 'Util/LocalizerUtil';
 import {sortUsersByPriority} from 'Util/StringUtil';
@@ -30,7 +31,6 @@ import {sortUsersByPriority} from 'Util/StringUtil';
 import {Conversation} from '../../../entity/Conversation';
 import {User} from '../../../entity/User';
 import {isServiceEntity} from '../../../guards/Service';
-import {SearchRepository} from '../../../search/SearchRepository';
 import {TeamRepository} from '../../../team/TeamRepository';
 import {PanelHeader} from '../PanelHeader';
 import {PanelEntity, PanelState} from '../RightSidebar';
