+++ src/script/repositories/conversation/ConversationRepository.test.ts
@@ -58,6 +58,7 @@ import {ClientEvent, CONVERSATION} from 'Repositories/event/Client';
 import {EventRepository} from 'Repositories/event/EventRepository';
 import {EventService} from 'Repositories/event/EventService';
 import {NOTIFICATION_HANDLING_STATE} from 'Repositories/event/NotificationHandlingState';
+import {SelfRepository} from 'Repositories/self/SelfRepository';
 import {Config} from 'src/script/Config';
 import {Conversation} from 'src/script/entity/Conversation';
 import {Message} from 'src/script/entity/message/Message';
@@ -85,7 +86,6 @@ import {NOTIFICATION_STATE} from './NotificationSetting';
 import {entities, payload} from '../../../../test/api/payloads';
 import {TestFactory} from '../../../../test/helper/TestFactory';
 import {generateUser} from '../../../../test/helper/UserGenerator';
-import {SelfRepository} from '../../self/SelfRepository';
 import {Core} from '../../service/CoreSingleton';
 import {LegacyEventRecord, StorageService} from '../../storage';
 import {TeamRepository} from '../../team/TeamRepository';
