+++ src/script/repositories/conversation/ConversationRepository.ts
@@ -72,6 +72,7 @@ import {EventService} from 'Repositories/event/EventService';
 import {EventSource} from 'Repositories/event/EventSource';
 import {NOTIFICATION_HANDLING_STATE} from 'Repositories/event/NotificationHandlingState';
 import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
+import {SelfRepository} from 'Repositories/self/SelfRepository';
 import {getNextItem} from 'Util/ArrayUtil';
 import {allowsAllFiles, getFileExtensionOrName, isAllowedFile} from 'Util/FileTypeUtil';
 import {replaceLink, t} from 'Util/LocalizerUtil';
@@ -153,7 +154,6 @@ import * as LegalHoldEvaluator from '../../legal-hold/LegalHoldEvaluator';
 import {MessageCategory} from '../../message/MessageCategory';
 import {SystemMessageType} from '../../message/SystemMessageType';
 import {addOtherSelfClientsToMLSConversation} from '../../mls';
-import {SelfRepository} from '../../self/SelfRepository';
 import {Core} from '../../service/CoreSingleton';
 import type {EventRecord} from '../../storage';
 import {ConversationRecord} from '../../storage/record/ConversationRecord';
