+++ src/script/repositories/search/SearchRepository.test.ts
@@ -23,11 +23,11 @@ import {createUuid} from 'Util/uuid';
 
 import {SearchRepository} from './SearchRepository';
 
-import {randomInt} from '../auth/util/randomUtil';
-import {generateUsers} from '../auth/util/test/TestUtil';
-import {APIClient} from '../service/APIClientSingleton';
-import {Core} from '../service/CoreSingleton';
-import {UserRepository} from '../user/UserRepository';
+import {randomInt} from '../../auth/util/randomUtil';
+import {generateUsers} from '../../auth/util/test/TestUtil';
+import {APIClient} from '../../service/APIClientSingleton';
+import {Core} from '../../service/CoreSingleton';
+import {UserRepository} from '../../user/UserRepository';
 
 function buildSearchRepository() {
   const userRepository = {getUsersById: jest.fn(() => [])} as unknown as jest.Mocked<UserRepository>;
