+++ src/script/user/UserRepository.test.ts
@@ -34,6 +34,7 @@ import {ClientMapper} from 'Repositories/client/ClientMapper';
 import {ConnectionEntity} from 'Repositories/connection/ConnectionEntity';
 import {EventRepository} from 'Repositories/event/EventRepository';
 import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
+import {SelfService} from 'Repositories/self/SelfService';
 import {entities} from 'test/api/payloads';
 import {TestFactory} from 'test/helper/TestFactory';
 import {generateAPIUser} from 'test/helper/UserGenerator';
@@ -45,7 +46,6 @@ import {UserService} from './UserService';
 import {UserState} from './UserState';
 
 import {User} from '../entity/User';
-import {SelfService} from '../self/SelfService';
 import {TeamState} from '../team/TeamState';
 import {serverTimeHandler} from '../time/serverTimeHandler';
