+++ src/script/user/UserRepository.ts
@@ -57,6 +57,7 @@ import {EventRepository} from 'Repositories/event/EventRepository';
 import type {EventSource} from 'Repositories/event/EventSource';
 import type {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
 import {PROPERTIES_TYPE} from 'Repositories/properties/PropertiesType';
+import type {SelfService} from 'Repositories/self/SelfService';
 import {chunk, partition} from 'Util/ArrayUtil';
 import {t} from 'Util/LocalizerUtil';
 import {getLogger, Logger} from 'Util/Logger';
@@ -75,7 +76,6 @@ import {Config} from '../Config';
 import {Conversation} from '../entity/Conversation';
 import {User} from '../entity/User';
 import {UserError} from '../error/UserError';
-import type {SelfService} from '../self/SelfService';
 import {UserRecord} from '../storage';
 import {TeamState} from '../team/TeamState';
 import type {ServerTimeHandler} from '../time/serverTimeHandler';
