+++ src/script/view_model/ActionsViewModel.ts
@@ -33,6 +33,7 @@ import type {MessageRepository} from 'Repositories/conversation/MessageRepositor
 import {NOTIFICATION_STATE} from 'Repositories/conversation/NotificationSetting';
 import type {IntegrationRepository} from 'Repositories/integration/IntegrationRepository';
 import type {ServiceEntity} from 'Repositories/integration/ServiceEntity';
+import {SelfRepository} from 'Repositories/self/SelfRepository';
 import {t} from 'Util/LocalizerUtil';
 import {isBackendError} from 'Util/TypePredicateUtil';
 
@@ -41,7 +42,6 @@ import type {MainViewModel} from './MainViewModel';
 import type {Conversation} from '../entity/Conversation';
 import type {Message} from '../entity/message/Message';
 import type {User} from '../entity/User';
-import {SelfRepository} from '../self/SelfRepository';
 import {UserState} from '../user/UserState';
 
 export class ActionsViewModel {
