+++ src/script/view_model/MainViewModel.ts
@@ -36,14 +36,14 @@ import type {NotificationRepository} from 'Repositories/notification/Notificatio
 import type {PreferenceNotificationRepository} from 'Repositories/notification/PreferenceNotificationRepository';
 import type {PermissionRepository} from 'Repositories/permission/PermissionRepository';
 import type {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
+import type {SearchRepository} from 'Repositories/search/SearchRepository';
+import type {SelfRepository} from 'Repositories/self/SelfRepository';
 
 import {ActionsViewModel} from './ActionsViewModel';
 import {CallingViewModel} from './CallingViewModel';
 import {ContentViewModel} from './ContentViewModel';
 import {ListViewModel} from './ListViewModel';
 
-import type {SearchRepository} from '../search/SearchRepository';
-import type {SelfRepository} from '../self/SelfRepository';
 import {Core} from '../service/CoreSingleton';
 import type {StorageRepository} from '../storage';
 import type {TeamRepository} from '../team/TeamRepository';
